requirements.txt
streamlit>=1.40.0
pandas>=2.2.0
mnemonic>=0.21
eth-utils>=5.0.0
base58>=2.1.1
pynacl>=1.5.0
pycryptodome>=3.20.0
ecdsa>=0.19.0